+++ bot.js
@@ -1,37 +1,18 @@
 require("dotenv").config()
-const { getArgs, textMenu } = require("./daakuu/utils")
+
+const { getArgs, textMenu } = require("./src/utils")
+const { closeDesc, readDesc, mainMenu } = require("./src/features/etc")
+const { reviewByUsers, detailReviewByUser } = require("./src/features/review")
+const { suggestions, detailSuggestion } = require("./src/features/recomendations")
+const { trendingManga, allTimePopulerManga, popularManhwa, top50Manga, searchManga, detailManga } = require("./src/features/manga")
+const  { trendingAnime, popularAnime, upcomingAnime, allTimePopulerAnime, top50Anime, topMovies, searchAnime, detailAnime } = require("./src/features/anime")
+
 const { Telegraf } = require("telegraf")
-
 const express = require('express')
-const app = express()
 
+const app = express()
 const port = process.env.PORT
 const bot = new Telegraf(process.env.BOT_TOKEN)
-
-const { 	
-	trendingAnime, 
-	popularAnime, 
-	upcomingAnime,
-	allTimePopulerAnime,
-	top50Anime,
-	topMovies,
-	trendingManga,
-	allTimePopulerManga,
-	popularManhwa,
-	top50Manga,
-	searchAnime,
-	searchManga,
-	suggestions,
-	reviewByUsers,
-	detailReviewByUser,
-	detailSuggestion,
-	detailAnime,
-	detailManga,
-	closeDesc,
-	readDesc,
-	mainMenu,
-} = require("./daakuu/features")
-
 
 bot.command(["/start", "/menu"], (ctx) => {
     ctx.reply(textMenu())
@@ -105,7 +86,7 @@ bot.command(["/about"], (ctx) => {
 	ctx.replyWithMarkdown("DaAkuu (Daftar Anime Kuu) is unofficial anilis.co bot.\n\nYou can visit our repository on GitHub to contribute or submit issue.",{ 
 		disable_web_page_preview: true,
 		"reply_markup":{
-			"inline_keyboard": [[{"text":"Open GitHub", "url": "https://github.com/Shiyinq/daakuu"}]]
+			"inline_keyboard": [[{"text":"Open GitHub", "url": "https://github.com/Shiyinq/src"}]]
 		}
 	})
 })
@@ -122,7 +103,7 @@ bot.launch()
 app.get('/', (_, res) => {
 	res.send('Bipp bipp!')
 })
-  
+
 app.listen(port, () => {
 	console.log(`Listening on port ${port}`)
 })
